Reorganized CodeLenses and improved empty files tests
