+++ packages/serverless-workflow-language-service/tests/SwfJsonLanguageService.test.ts
@@ -18,8 +18,9 @@ import {
   JsonCodeCompletionStrategy,
   SwfJsonLanguageService,
 } from "@kie-tools/serverless-workflow-language-service/dist/channel";
+import * as path from "path";
+import * as fs from "fs";
 import { CodeLens, CompletionItem, CompletionItemKind, InsertTextFormat, Position } from "vscode-languageserver-types";
-import * as simpleTemplate from "@kie-tools/serverless-workflow-language-service/dist/assets/code-completion/simple-template.sw.json";
 import {
   defaultConfig,
   defaultServiceCatalogConfig,
@@ -28,6 +29,7 @@ import {
 } from "./SwfLanguageServiceConfigs";
 import { codeCompletionTester, ContentWithCursor, getStartNodeValuePositionTester, trim } from "./testUtils";
 
+const EXPECTED_RESULTS_PROJECT_FOLDER: string = path.resolve("tests", "expectedResults");
 const documentUri = "test.sw.json";
 
 describe("JsonCodeCompletionStrategy", () => {
@@ -458,7 +460,11 @@ describe("SWF LS JSON", () => {
         ["empty file with a newline before the cursor", `\n🎯`],
         ["empty file with a newline after the cursor", `🎯\n`],
       ])("%s", async (_description, content: ContentWithCursor) => {
-        let { completionItems, cursorPosition } = await codeCompletionTester(ls, documentUri, content, false);
+        const { completionItems, cursorPosition } = await codeCompletionTester(ls, documentUri, content, false);
+        const expectedResult = fs.readFileSync(
+          path.resolve(EXPECTED_RESULTS_PROJECT_FOLDER, "emptyfile_autocompletion.sw.json.result"),
+          "utf-8"
+        );
 
         expect(completionItems).toHaveLength(1);
         expect(completionItems[0]).toStrictEqual({
@@ -468,7 +474,7 @@ describe("SWF LS JSON", () => {
           detail: "Start with a simple Serverless Workflow",
           textEdit: {
             range: { start: cursorPosition, end: cursorPosition },
-            newText: JSON.stringify(simpleTemplate, null, 2),
+            newText: expectedResult,
           },
           insertTextFormat: InsertTextFormat.Snippet,
         } as CompletionItem);
