+++ packages/serverless-workflow-language-service/tests/SwfYamlLanguageService.test.ts
@@ -23,7 +23,8 @@ import {
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { CodeLens, CompletionItem, CompletionItemKind, InsertTextFormat, Position } from "vscode-languageserver-types";
 import { dump } from "yaml-language-server-parser";
-import * as simpleTemplate from "@kie-tools/serverless-workflow-language-service/dist/assets/code-completion/simple-template.sw.json";
+import * as path from "path";
+import * as fs from "fs";
 import {
   defaultConfig,
   defaultServiceCatalogConfig,
@@ -32,6 +33,7 @@ import {
 } from "./SwfLanguageServiceConfigs";
 import { codeCompletionTester, ContentWithCursor, getStartNodeValuePositionTester, treat, trim } from "./testUtils";
 
+const EXPECTED_RESULTS_PROJECT_FOLDER: string = path.resolve("tests", "expectedResults");
 const documentUri = "test.sw.yaml";
 
 describe("YamlCodeCompletionStrategy", () => {
@@ -913,6 +915,10 @@ functions:
         ["empty file with a newline after the cursor", `🎯\n`],
       ])("%s", async (_description, content: ContentWithCursor) => {
         let { completionItems, cursorPosition } = await codeCompletionTester(ls, documentUri, content, false);
+        const expectedResult = fs.readFileSync(
+          path.resolve(EXPECTED_RESULTS_PROJECT_FOLDER, "emptyfile_autocompletion.sw.yaml.result"),
+          "utf-8"
+        );
 
         expect(completionItems).toHaveLength(1);
         expect(completionItems[0]).toStrictEqual({
@@ -922,7 +928,7 @@ functions:
           detail: "Start with a simple Serverless Workflow",
           textEdit: {
             range: { start: cursorPosition, end: cursorPosition },
-            newText: dump(simpleTemplate, {}).slice(0, -1),
+            newText: expectedResult,
           },
           insertTextFormat: InsertTextFormat.Snippet,
         } as CompletionItem);
