+++ packages/serverless-workflow-language-service/tests/expectedResults/emptyfile_autocompletion.sw.yaml.result
@@ -1,0 +1,25 @@
+id: Workflow unique identifier
+version: '0.1'
+specVersion: '0.1'
+name: Workflow name
+description: Workflow description
+start: StartState
+functions:
+  - name: uniqueFunctionName
+    operation: 'localhost#operation'
+    type: rest
+events:
+  - name: Unique event name
+    source: CloudEvent source
+    type: CloudEvent type
+states:
+  - name: StartState
+    type: operation
+    actions:
+      - name: uniqueActionName
+        functionRef:
+          refName: uniqueFunctionName
+          arguments:
+            firstArgument: ''
+            secondArgument: ''
+    end: true
